requirements.txt
cffi==1.17.1
cloudpickle==3.1.0
filelock==3.16.1
fsspec==2024.2.0
glcontext==3.0.0
glfw==1.12.0
gym==0.26.2
gym-notices==0.0.8
gym3==0.3.3
imageio==2.36.1
imageio-ffmpeg==0.3.0
importlib_metadata==8.5.0
Jinja2==3.1.3
MarkupSafe==2.1.5
moderngl==5.12.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
pillow==11.0.0
procgen==0.10.7
pycparser==2.22
pygame==2.6.1
sympy==1.13.1
torch==2.5.1+cu121
torchvision==0.20.1+cu121
typing_extensions==4.9.0
zipp==3.21.0
